Improve backend layout
